fix($animate): ensure $animate doesn't break natural CSS transitions
BREAKING CHANGE

ngClass and {{ class }} will call the setClass animation callback instead
of addClass / removeClass when both a addClass/removeClass operation is
being executed on the element during the animation. Please include
the setClass animation callback as well as addClass and removeClass within
your JS animations to work with ngClass and {{ class }} directives.

Closes #6019